requirements.txt
absl-py==1.4.0
Adafruit-Blinka==8.18.0
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.0
adafruit-io==2.7.0
Adafruit-PlatformDetect==3.44.0
Adafruit-PureIO==1.1.10
argparse
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
flatbuffers==20181003210633
idna==3.4
numpy==1.24.1
opencv-python==4.5.3.56
paho-mqtt==1.6.1
Pillow==9.5.0
pkg_resources==0.0.0
protobuf==3.20.3
pybind11==2.10.3
pycparser==2.21
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
requests==2.28.2
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
sounddevice==0.4.5
sysv-ipc==1.1.0
tflite-support==0.4.3
thread6==0.2.0
typing_extensions==4.5.0
urllib3==1.26.15
